+++ Fubuki/FFmpeg/BGR24Handler.h
@@ -23,7 +23,7 @@ namespace OnePunchMan
 		* @param: height 图片高度
 		* @param: packetIndex 帧序号
 		*/
-		void HandleFrame(unsigned char* bgr24, int width, int height, long long packetIndex);
+		void HandleFrame(unsigned char* bgr24, int width, int height, int packetIndex);
 
 	private:
 		//当前写入数量
